考试管理基础功能

+++ ruoyi-modules/ruoyi-exam/src/main/java/com/ruoyi/exam/service/impl/QuestionServiceImpl.java
@@ -7,7 +7,15 @@ import com.ruoyi.common.mybatis.core.page.TableDataInfo;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
-import com.ruoyi.exam.domain.ExamPaper;
+import com.ruoyi.exam.contants.ExamStatusConstants;
+import com.ruoyi.exam.domain.ExamSend;
+import com.ruoyi.exam.domain.enums.QuestionTypeEnum;
+import com.ruoyi.exam.domain.vo.ExamPaperVo;
+import com.ruoyi.exam.domain.vo.ExamSendVo;
+import com.ruoyi.exam.exception.BusinessException;
+import com.ruoyi.exam.exception.ErrorCode;
+import com.ruoyi.exam.mapper.ExamPaperMapper;
+import com.ruoyi.exam.mapper.ExamSendMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import com.ruoyi.exam.domain.bo.QuestionBo;
@@ -15,28 +23,32 @@ import com.ruoyi.exam.domain.vo.QuestionVo;
 import com.ruoyi.exam.domain.Question;
 import com.ruoyi.exam.mapper.QuestionMapper;
 import com.ruoyi.exam.service.IQuestionService;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Collection;
+import java.util.stream.Collectors;
 
 /**
  * 题目Service业务层处理
  *
  * @author zkm
- * @date  2025-03
+ * @date 2025-03
  */
 @RequiredArgsConstructor
 @Service
 public class QuestionServiceImpl implements IQuestionService {
 
     private final QuestionMapper baseMapper;
+    private final ExamPaperMapper paperMapper;
+    private final ExamSendMapper sendMapper;
 
     /**
      * 查询题目
      */
     @Override
-    public QuestionVo queryById(Long id){
+    public QuestionVo queryById(Long id) {
         return baseMapper.selectVoById(id);
     }
 
@@ -65,6 +77,7 @@ public class QuestionServiceImpl implements IQuestionService {
         lqw.eq(bo.getQuestionTypeCode() != null, Question::getQuestionTypeCode, bo.getQuestionTypeCode());
         lqw.like(StringUtils.isNotBlank(bo.getQuestionText()), Question::getQuestionText, bo.getQuestionText());
         lqw.eq(bo.getScore() != null, Question::getScore, bo.getScore());
+        lqw.orderBy(true, false, Question::getUpdateTime);
         lqw.orderBy(true, false, Question::getCreateTime);
         return lqw;
     }
@@ -74,8 +87,8 @@ public class QuestionServiceImpl implements IQuestionService {
      */
     @Override
     public Boolean insertByBo(QuestionBo bo) {
-        Question add = BeanUtil.toBean(bo, Question.class);
-        validEntityBeforeSave(add);
+
+        Question add = buildEntity(bo);
         boolean flag = baseMapper.insert(add) > 0;
         if (flag) {
             bo.setId(add.getId());
@@ -83,31 +96,73 @@ public class QuestionServiceImpl implements IQuestionService {
         return flag;
     }
 
+    // 构造实体对象
+    private Question buildEntity(QuestionBo bo) {
+        Question entity = BeanUtil.copyProperties(bo, Question.class);
+        // 校验题型
+        Integer typeCode = entity.getQuestionTypeCode();
+        QuestionTypeEnum.fromCode(typeCode);
+        return entity;
+    }
+
     /**
      * 修改题目
      */
     @Override
+    @Transactional
     public Boolean updateByBo(QuestionBo bo) {
-        Question update = BeanUtil.toBean(bo, Question.class);
-        validEntityBeforeSave(update);
-        return baseMapper.updateById(update) > 0;
+        Question update = buildEntity(bo);
+
+        // 查询题目是否被试卷引用
+        List<ExamPaperVo> relatedPapers = paperMapper.selectByQuestionId(bo.getId());
+
+        if (relatedPapers.isEmpty()) {
+            // 题目未被试卷引用，直接修改
+            return baseMapper.updateById(update) > 0;
+        }
+
+        // 考试未结束
+        boolean hasOngoingExam = isHasOngoingExam(relatedPapers);
+        if (hasOngoingExam) {
+            throw new BusinessException(ErrorCode.OPERATION_ERROR, "当前题目正在考试中，无法修改！");
+        }
+
+        // 考试已结束，逻辑删除旧题目，新增新题目
+        baseMapper.deleteById(bo.getId());
+        update.setId(null);
+        return baseMapper.insert(update) > 0;
     }
 
-    /**
-     * 保存前的数据校验
-     */
-    private void validEntityBeforeSave(Question entity){
-        //TODO 做一些数据校验,如唯一约束
-    }
 
     /**
      * 批量删除题目
      */
     @Override
     public Boolean deleteWithValidByIds(Collection<Long> ids, Boolean isValid) {
-        if(isValid){
-            //TODO 做一些业务上的校验,判断是否需要校验
+
+        // 查询题目是否被试卷引用
+        List<ExamPaperVo> relatedPapers = paperMapper.selectByQuestionIds(ids);
+
+        // 判断是否在考试中
+        boolean hasOngoingExam = !relatedPapers.isEmpty() && isHasOngoingExam(relatedPapers);
+        if (hasOngoingExam) {
+            throw new BusinessException(ErrorCode.OPERATION_ERROR, "题目正在考试中，无法删除！");
         }
+
         return baseMapper.deleteBatchIds(ids) > 0;
+    }
+
+    /**
+     * 判断发放考试是否结束
+     * @param relatedPapers 试卷信息
+     * @return true 考试中
+     */
+    private boolean isHasOngoingExam(List<ExamPaperVo> relatedPapers) {
+        List<Long> collect = relatedPapers.stream().map(ExamPaperVo::getId).collect(Collectors.toList());
+
+        List<ExamSendVo> sendVoList = sendMapper.selectVoList(Wrappers.<ExamSend>lambdaQuery()
+            .in(ExamSend::getPaperId, collect));
+        return sendVoList.stream()
+            .anyMatch(send -> send.getStatus() != ExamStatusConstants.STATUS_COMPLETED);
     }
 }
